requirements.txt
appengine-python-standard>=1.0.0
Flask==2.2.2
google-cloud-datastore==2.7.1
requests==2.28.1
yt-dlp==2023.1.6
